+++ sdk/resourcemanager/consumption/armconsumption/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Release History
 
+## 2.0.0 (2023-05-11)
+### Breaking Changes
+
+- Function `*ReservationRecommendationDetailsClient.Get` parameter(s) have been changed from `(context.Context, string, Scope, string, Term, LookBackPeriod, string, *ReservationRecommendationDetailsClientGetOptions)` to `(context.Context, string, string, *ReservationRecommendationDetailsClientGetOptions)`
+- Type of `ModernReservationRecommendation.Properties` has been changed from `*ModernReservationRecommendationProperties` to `ModernReservationRecommendationPropertiesClassification`
+- Enum `LookBackPeriod` has been removed
+- Enum `Scope` has been removed
+- Enum `Term` has been removed
+- Field `Etag`, `Tags` of struct `CreditSummary` has been removed
+- Field `MarketplaceCharges` of struct `LegacyChargeSummaryProperties` has been removed
+
+### Features Added
+
+- New value `EventTypeCreditExpired` added to enum type `EventType`
+- New function `*ModernReservationRecommendationProperties.GetModernReservationRecommendationProperties() *ModernReservationRecommendationProperties`
+- New function `*ModernSharedScopeReservationRecommendationProperties.GetModernReservationRecommendationProperties() *ModernReservationRecommendationProperties`
+- New function `*ModernSingleScopeReservationRecommendationProperties.GetModernReservationRecommendationProperties() *ModernReservationRecommendationProperties`
+- New struct `ModernSharedScopeReservationRecommendationProperties`
+- New struct `ModernSingleScopeReservationRecommendationProperties`
+- New field `OverageRefund` in struct `BalanceProperties`
+- New field `ETag` in struct `CreditSummary`
+- New field `IsEstimatedBalance` in struct `CreditSummaryProperties`
+- New field `IsEstimatedBalance` in struct `EventProperties`
+- New field `AzureMarketplaceCharges` in struct `LegacyChargeSummaryProperties`
+- New field `IsEstimatedBalance` in struct `LotProperties`
+- New field `SubscriptionID` in struct `ModernChargeSummaryProperties`
+- New field `PreviewMarkupPercentage`, `UseMarkupIfPartner` in struct `ReservationTransactionsClientListOptions`
+
+
 ## 1.1.0 (2023-03-28)
 ### Features Added